+++ packages/at_cli_commons/lib/src/cli_base.dart
@@ -1,8 +1,8 @@
-
 import 'dart:io';
 
 import 'package:args/args.dart';
 import 'package:at_cli_commons/src/service_factories.dart';
+import 'package:at_cli_commons/src/utils.dart';
 import 'package:at_client/at_client.dart';
 import 'package:at_onboarding_cli/at_onboarding_cli.dart';
 import 'package:at_utils/at_logger.dart';
@@ -11,28 +11,70 @@ import 'package:logging/logging.dart';
 import 'package:version/version.dart';
 
 class CLIBase {
+  /// An ArgParser which has all of the options and flags required by [CLIBase]
+  /// Used by [fromCommandLineArgs] if the `parser` parameter isn't supplied.
   static final ArgParser argsParser = ArgParser()
-    ..addFlag('help', negatable: false,
-        help: 'Usage instructions')
-    ..addOption('atsign', abbr: 'a', mandatory: true,
-        help: 'This client\'s atSign')
-    ..addOption('namespace', abbr: 'n', mandatory: false,
-        help: 'Namespace')
-    ..addOption('key-file', abbr: 'k', mandatory: false,
+    ..addFlag('help', negatable: false, help: 'Usage instructions')
+    ..addOption('atsign',
+        abbr: 'a', mandatory: true, help: 'This client\'s atSign')
+    ..addOption('namespace', abbr: 'n', mandatory: true, help: 'Namespace')
+    ..addOption('key-file',
+        abbr: 'k',
+        mandatory: false,
         help: 'Your atSign\'s atKeys file if not in ~/.atsign/keys/')
-    ..addOption('cram-secret', abbr: 'c', mandatory: false,
-        help: 'atSign\'s cram secret')
-    ..addOption('home-dir', abbr: 'h', mandatory: false,
-        help: 'home directory')
-    ..addOption('storage-dir', abbr: 's', mandatory: false,
+    ..addOption('cram-secret',
+        abbr: 'c', mandatory: false, help: 'atSign\'s cram secret')
+    ..addOption('home-dir', abbr: 'h', mandatory: false, help: 'home directory')
+    ..addOption('storage-dir',
+        abbr: 's',
+        mandatory: false,
         help: 'directory for this client\'s local storage files')
-    ..addOption('root-domain', abbr: 'd', mandatory: false,
-        help: 'Root Domain', defaultsTo: 'root.atsign.org')
-    ..addFlag('verbose', abbr: 'v', negatable: false,
-        help: 'More logging')
-    ..addFlag('never-sync', negatable: false,
-        help: 'Do not run sync')
-  ;
+    ..addOption('root-domain',
+        abbr: 'd',
+        mandatory: false,
+        help: 'Root Domain',
+        defaultsTo: 'root.atsign.org')
+    ..addFlag('verbose', abbr: 'v', negatable: false, help: 'More logging')
+    ..addFlag('never-sync', negatable: false, help: 'Do not run sync');
+
+  /// Constructs a CLIBase from a list of command-line arguments
+  /// and calls [init] on it.
+  /// <br/>
+  /// <br/>
+  /// If [parser] is not supplied then uses CLIBase's [argsParser] static var.
+  /// Allowing [parser] to be supplied enables callers to do something like this:
+  /// ```
+  ///     ArgParser argsParser = CLIBase.argsParser
+  ///       ..addOption('my-cli-option',
+  ///          help: "an option which configures my cli's unique feature");
+  ///
+  ///     CLIBase cliBase = await CLIBase.fromCommandLineArgs(args, parser: argsParser);
+  /// ```
+  static Future<CLIBase> fromCommandLineArgs(List<String> args,
+      {ArgParser? parser}) async {
+    parser ??= argsParser;
+    ArgResults parsedArgs = parser.parse(args);
+
+    if (parsedArgs['help'] == true) {
+      print(parser.usage);
+      exit(0);
+    }
+
+    CLIBase cliBase = CLIBase(
+        atSign: parsedArgs['atsign'],
+        atKeysFilePath: parsedArgs['key-file'],
+        nameSpace: parsedArgs['namespace'],
+        rootDomain: parsedArgs['root-domain'],
+        homeDir: getHomeDirectory(),
+        storageDir: parsedArgs['storage-dir'],
+        verbose: parsedArgs['verbose'],
+        cramSecret: parsedArgs['cram-secret'],
+        syncDisabled: parsedArgs['never-sync']);
+
+    await cliBase.init();
+
+    return cliBase;
+  }
 
   final String atSign;
   final String nameSpace;
@@ -53,34 +95,55 @@ class CLIBase {
   late final AtSignLogger logger;
   late final AtClient atClient;
 
-  CLIBase({
-    required this.atSign,
-    required this.nameSpace,
-    required this.rootDomain,
-    this.homeDir,
-    this.verbose=false,
-    this.atKeysFilePath,
-    this.storageDir,
-    this.downloadDir,
-    this.cramSecret,
-    this.syncDisabled = false}) {
-
+  /// Validates parameters and constructs a CLIBase instance.
+  /// <br/> <br/>
+  /// Validation rules:
+  /// - homeDir must be non-null when any of the atKeysFilePath, storageDir or
+  ///   downloadDir parameters are null
+  ///
+  /// <br/>
+  /// Also configures the default AtSignLogger log level to be either INFO
+  /// if verbose is true, or SHOUT if verbose is false (the default). If the
+  /// application wishes to use a different default log level then it can do
+  /// something like this:
+  /// ```
+  ///     AtSignLogger.root_level = 'FINEST';
+  ///     cliBase.logger.logger.level = Level.FINEST;
+  /// ```
+  /// Throws an [IllegalArgumentException] if the parameters fail validation.
+  CLIBase(
+      {required this.atSign,
+      required this.nameSpace,
+      required this.rootDomain,
+      this.homeDir,
+      this.verbose = false,
+      this.atKeysFilePath,
+      this.storageDir,
+      this.downloadDir,
+      this.cramSecret,
+      this.syncDisabled = false}) {
     if (homeDir == null) {
       if (atKeysFilePath == null) {
-        throw IllegalArgumentException('homeDir must be provided when atKeysFilePath is not provided');
+        throw IllegalArgumentException(
+            'homeDir must be provided when atKeysFilePath is not provided');
       }
       if (storageDir == null) {
-        throw IllegalArgumentException('homeDir must be provided when storageDir is not provided');
+        throw IllegalArgumentException(
+            'homeDir must be provided when storageDir is not provided');
       }
       if (downloadDir == null) {
-        throw IllegalArgumentException('homeDir must be provided when downloadDir is not provided');
+        throw IllegalArgumentException(
+            'homeDir must be provided when downloadDir is not provided');
       }
     }
 
-    atKeysFilePathToUse = atKeysFilePath ?? '$homeDir/.atsign/keys/${atSign}_key.atKeys';
-    localStoragePathToUse = storageDir ?? '$homeDir/.$nameSpace/$atSign/storage';
+    atKeysFilePathToUse =
+        atKeysFilePath ?? '$homeDir/.atsign/keys/${atSign}_key.atKeys';
+    localStoragePathToUse =
+        storageDir ?? '$homeDir/.$nameSpace/$atSign/storage';
     downloadPathToUse = downloadDir ?? '$homeDir/.$nameSpace/$atSign/files';
 
+    AtSignLogger.defaultLoggingHandler = AtSignLogger.stdErrLoggingHandler;
 
     logger = AtSignLogger(runtimeType.toString());
     logger.hierarchicalLoggingEnabled = true;
@@ -89,11 +152,11 @@ class CLIBase {
       logger.logger.level = Level.INFO;
     } else {
       AtSignLogger.root_level = 'SHOUT';
-      // logger.logger.level = Level.SHOUT;
-      logger.logger.level = Level.INFO;
+      logger.logger.level = Level.SHOUT;
     }
   }
 
+  /// Does the various things required to create an AtClient object
   Future<void> init() async {
     AtServiceFactory? atServiceFactory;
 
@@ -114,9 +177,11 @@ class CLIBase {
       ..cramSecret = cramSecret
       ..atProtocolEmitted = Version(2, 0, 0);
 
-    AtOnboardingService onboardingService = AtOnboardingServiceImpl(atSign, atOnboardingConfig, atServiceFactory: atServiceFactory);
+    AtOnboardingService onboardingService = AtOnboardingServiceImpl(
+        atSign, atOnboardingConfig,
+        atServiceFactory: atServiceFactory);
 
-    if (! File(atKeysFilePathToUse).existsSync()) {
+    if (!File(atKeysFilePathToUse).existsSync()) {
       await onboardingService.onboard();
     }
 
@@ -125,10 +190,13 @@ class CLIBase {
     while (!authenticated) {
       try {
         stdout.write(chalk.brightBlue('\r\x1b[KConnecting ... '));
-        await Future.delayed(Duration(milliseconds: 1000)); // Pause just long enough for the retry to be visible
+        await Future.delayed(Duration(
+            milliseconds:
+                1000)); // Pause just long enough for the retry to be visible
         authenticated = await onboardingService.authenticate();
       } catch (exception) {
-        stdout.write(chalk.brightRed('$exception. Will retry in ${retryDuration.inSeconds} seconds'));
+        stdout.write(chalk.brightRed(
+            '$exception. Will retry in ${retryDuration.inSeconds} seconds'));
       }
       if (!authenticated) {
         await Future.delayed(retryDuration);