+++ packages/at_cli_commons/lib/src/pipe_print.dart
@@ -1,8 +1,0 @@
-import 'dart:io';
-import 'package:chalkdart/chalk.dart';
-
-void pipePrint (String p){
-  if (stdin.hasTerminal){
-    stdout.write(chalk.red.bold(p));
-  }
-}