+++ packages/at_commons/lib/at_commons.dart
@@ -35,5 +35,6 @@ export 'package:at_commons/src/verb/syntax.dart';
 export 'package:at_commons/src/verb/update_json.dart';
 export 'package:at_commons/src/verb/verb_util.dart';
 export 'package:at_commons/src/auth/auth_mode.dart';
+export 'package:at_commons/src/verb/enroll_params.dart';
 @experimental
 export 'package:at_commons/src/telemetry/at_telemetry.dart';