+++ packages/at_cli_commons/CHANGELOG.md
@@ -1,2 +1,11 @@
+## 1.0.3
+- Added `example/` package, moved code samples from `bin/` to `example/`
+## 1.0.2
+- docs: Added some code samples in bin/ directory
+- docs: Added some class and method documentation to CLIBase
+- docs: Updated README
+- feat: Added static `fromCommandLineArgs` factory method to CLIBase
+## 1.0.1
+- Small edits to README
 ## 1.0.0
 - Initial version.